Refactor pyobd
