+++ src/pyobd/gui/about_text.py
@@ -1,0 +1,20 @@
+ABOUT_TEXT = """  PyOBD is an automotive OBD2 diagnosting application using ELM237 cable.
+
+(C) 2021 Jure Poljsak
+(C) 2008-2009 SeCons Ltd.
+(C) 2004 Charles Donour Sizemore
+
+https://github.com/barracuda-fsh/pyobd
+http://www.obdtester.com/
+http://www.secons.com/
+
+  PyOBD is free software; you can redistribute it and/or modify
+it under the terms of the GNU General Public License as published by the Free Software Foundation; 
+either version 2 of the License, or (at your option) any later version.
+
+  PyOBD is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; 
+without even the implied warranty of MEHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  
+See the GNU General Public License for more details. You should have received a copy of 
+the GNU General Public License along with PyOBD; if not, write to 
+the Free Software Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
+"""
